♻️ refactor: Remove Scramble component and add scramble mode to TextTicker
- Remove Scramble component, hooks, styles, and related exports
- Add `scrambleDuration` and `staggerDelay` props to TextTicker for deterministic scramble mode
- Update docs, demos, and styles‑API to reflect new behavior and remove Scramble references
