--- package/src/Scramble/Scramble.module.css
@@ -1,8 +1,0 @@
-.root {
-  transition-property: color, background-color, border-color;
-  transition-timing-function: ease;
-  transition-duration: 0.2s;
-  display: inline-block;
-  font-family: monospace;
-  white-space: pre;
-}
