+++ package/src/TextAnimate.tsx
@@ -20,7 +20,6 @@ import { Gradient } from './Gradient/Gradient';
 import { Highlight } from './Highlight/Highlight';
 import { Morphing } from './Morphing/Morphing';
 import { NumberTicker } from './NumberTicker/NumberTicker';
-import { Scramble } from './Scramble/Scramble';
 import { Spinner } from './Spinner/Spinner';
 import { SplitFlap } from './SplitFlap/SplitFlap';
 import { TextTicker } from './TextTicker/TextTicker';
@@ -248,7 +247,6 @@ export type TextAnimateFactory = PolymorphicFactory<{
     TextTicker: typeof TextTicker;
     Gradient: typeof Gradient;
     Highlight: typeof Highlight;
-    Scramble: typeof Scramble;
     SplitFlap: typeof SplitFlap;
     Morphing: typeof Morphing;
   };
@@ -504,6 +502,5 @@ TextAnimate.NumberTicker = NumberTicker;
 TextAnimate.TextTicker = TextTicker;
 TextAnimate.Gradient = Gradient;
 TextAnimate.Highlight = Highlight;
-TextAnimate.Scramble = Scramble;
 TextAnimate.SplitFlap = SplitFlap;
 TextAnimate.Morphing = Morphing;
