+++ package/src/index.ts
@@ -19,7 +19,6 @@ export type { GradientFactory, GradientProps } from './Gradient/Gradient';
 export type { HighlightFactory, HighlightProps } from './Highlight/Highlight';
 export type { MorphingFactory, MorphingProps } from './Morphing/Morphing';
 export type { NumberTickerFactory, NumberTickerProps } from './NumberTicker/NumberTicker';
-export type { ScrambleFactory, ScrambleProps } from './Scramble/Scramble';
 export type { SpinnerFactory, SpinnerProps } from './Spinner/Spinner';
 export type { SplitFlapFactory, SplitFlapProps } from './SplitFlap/SplitFlap';
 export type { TextTickerFactory, TextTickerProps } from './TextTicker/TextTicker';
@@ -27,7 +26,6 @@ export type { TypewriterFactory, TypewriterProps } from './Typewriter/Typewriter
 
 export { useMorphing } from './Morphing/use-morphing';
 export { useNumberTicker, type NumberTickerEasing } from './NumberTicker/use-number-ticker';
-export { useScramble } from './Scramble/use-scramble';
 export { useSplitFlap } from './SplitFlap/use-split-flap';
 export { useTextAnimate, type UseTextAnimateResult } from './use-text-animate';
 export { useTextTicker } from './TextTicker/use-text-ticker';
